+++ lib/dovico/model/client.rb
@@ -1,0 +1,4 @@
+module Dovico
+  class Client < Assignment
+  end
+end