+++ spec/helper.rb
@@ -15,6 +15,8 @@ require "rspec/its"
 require "timecop"
 require "rack/test"
 require 'webmock/rspec'
+require 'fileutils'
+require 'tmpdir'
 
 require 'pry'
 
@@ -23,3 +25,15 @@ require "dovico"
 
 Timecop.safe_mode = true
 WebMock.disable_net_connect!(allow_localhost: true)
+
+RSpec.configure do |config|
+  # Use a temporary directory for cache file
+  config.before(:example) do
+    @cache_dir = Dir.mktmpdir("dovico")
+    stub_const("Dovico::Assignments::CACHE_FILE", "#{@cache_dir}/assignments.json")
+  end
+
+  config.after(:example) do
+    FileUtils.rm_rf(@cache_dir)
+  end
+end